# Add vectorized h3_to_parent and h3_get_resolution