--- a/src/h3/_cy/h3lib.pxd
+++ b/src/h3/_cy/h3lib.pxd
@@ -88,7 +88,7 @@ cdef extern from "h3api.h":
 
     stdint.int64_t numHexagons(int res)
 
-    int h3GetResolution(H3Index h)
+    int h3GetResolution(H3Index h) nogil
 
     int h3GetBaseCell(H3Index h)