This PR adds a simple vectorized `h3_to_parent`, along the lines of #147 . I compared this to a simple loop that repeatedly calls `h3.h3_to_parent` and found a 160x speedup over pure Python and 100x faster than `np.vectorize`.

I'd love to add a vectorized complement to every h3 API function, but that would create a lot of duplication, as presumably every function would need a for loop in Cython, so you couldn't use the `_api_template` setup to add vectorized functions.

```py
import h3.api.numpy_int as h3
import h3.unstable.vect as vect
import numpy as np

N = 100_000
K = 10

lats, longs = np.random.uniform(0, 90, N), np.random.uniform(0, 90, N)
h3_index = vect.geo_to_h3(lats, longs, 9)

%%timeit
out_loop = []
for i in range(len(h3_index)):
    out_loop.append(h3.h3_to_parent(h3_index[i], 8))
# 67.6 ms ± 2.06 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

vec_fn = np.vectorize(h3.h3_to_parent)
%timeit vec_fn(h3_index, 8)
# 42.6 ms ± 1.23 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

%timeit out_vect = vect.h3_to_parent(h3_index, 8)
# 423 µs ± 15.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

np.array_equal(out_loop, out_vect)
# True
```

cc @isaacbrodsky