@ajfriend 2020-09-03
And one open question is what we do with invalid inputs for these vectorized functions. For the non-vectorized ones, we might raise an exception.

Errors might be:
- the input array contains an invalid h3 cell
- an input cell already has lower resolution than the given resolution.

For the vectorized ones, do we just return `H3_NULL`? Fail the whole function call?