@ajfriend 2020-09-03
Another question: Would we ever want to vectorize the resolution parameter in `3_to_parent`, or are we OK with forcing it to always be a single `int`?