@kylebarron 2020-09-03
Re exceptions, I think it would be fine to fail the whole function call, but open to suggestions. It might be good to look into how numpy handles exceptions.

As it is in this PR, it appears that invalid values are coerced to 0?
```py
vect.h3_to_parent(np.array([17584386400583679], dtype=np.uint64), 8)
# array([0], dtype=uint64)
```

Input arrays that aren't `np.uint64` will get a `ValueError`:
```
vect.h3_to_parent(np.array([1], dtype=np.uint8), 8)
---------------------------------------------------------------------------
ValueError                                Traceback (most recent call last)
<ipython-input-71-3751d4de5ad0> in <module>
----> 1 vect.h3_to_parent(np.array([1], dtype=np.uint8), 8)

~/local/anaconda3/lib/python3.7/site-packages/h3/unstable/vect.py in h3_to_parent(h, res)
     21     out = np.zeros(len(h), dtype='uint64')
     22 
---> 23     _vect.h3_to_parent_vect(h, res, out)
     24 
     25     return out

unstable_vect.pyx in h3._cy.unstable_vect.h3_to_parent_vect()

ValueError: Buffer dtype mismatch, expected 'const H3Index' but got 'unsigned char'
```


> Would we ever want to vectorize the resolution parameter in `h3_to_parent`, or are we OK with forcing it to always be a single `int`?

I think it's better to keep it an `int` for simplicity. Someone might care about a few resolution levels, but the number of hex values will dwarf the number of resolution values, so IMO better to vectorize over hex and leave resolution as `int`.