@kylebarron 2020-09-04
To sum up a bit more research: it's possible to iterate over arbitrary multi-dimensional arrays with the C-API. Example on my [`vectorize-ufunc`](https://github.com/kylebarron/h3-py/blob/075df7ada80128e23096ba84fccdc1d4ed52d3e2/src/h3/_cy/unstable_vect.pyx#L72-L94) branch.

It should also be possible [using `np.nditer`](https://stackoverflow.com/a/39058906), but that may be slower than the C-API.

Possibly the simplest approach, if the variety of inputs is less complex but you still want to accept arbitrary dimension inputs, you can use `np.ravel` to get a 1-dimensional view on n-dimensional input, and then `reshape` back to the original input. https://stackoverflow.com/a/33316468

Since `np.ravel` can return data in the same order, a function like `geoToH3` could take two arbitrary-dimension arrays for `lat` and `lon`, check that the shapes are the same, and then use `np.ravel` to iterate over the data and return an `H3Index` array of the same shape as the input `lat` and `lon`.

All of the above approaches require a dependency on Numpy at build time and runtime, so presumably there would be a flag from Python that only runs this code if the user has Numpy installed.