@kylebarron 2021-02-26
I was looking at this PR again and realized that it wasn't as far as I thought from merge-ready, so I figured I try to push to get it over the finish line.

**Final changes:**

- Add a default argument for `h3_to_parent`. If `res` isn't provided, then the parent one resolution up _from each cell_ is used. Note that this happens per-cell, so if you have an array of cells at resolution `[9, 8, 7]`, and `res` is `None`, then the results will be at resolution `[8, 7, 6]`.
- Allow `res` to `h3_to_parent` to be an array. This allows `res` to vary for each hex. I haven't tested broadcasting where the hex input and `res` are different shapes (I expect it would crash).
- Add a test file

Let me know if you have any more comments.

I also did the original benchmarks again and found similar results (except it looks like Python 3.8 has some `list` improvements 😄 )

```py
import h3.api.numpy_int as h3
import h3.unstable.vect as vect
import numpy as np

N = 100_000
K = 10

lats, longs = np.random.uniform(0, 90, N), np.random.uniform(0, 90, N)
h3_index = vect.geo_to_h3(lats, longs, 9)

%%timeit
out_loop = []
for i in range(len(h3_index)):
    out_loop.append(h3.h3_to_parent(h3_index[i], 8))
# 55.5 ms ± 1.66 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

vec_fn = np.vectorize(h3.h3_to_parent)
%timeit vec_fn(h3_index, 8)
# 41.2 ms ± 960 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

%timeit out_vect = vect.h3_to_parent(h3_index, 8)
# 411 µs ± 16.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

np.array_equal(out_loop, out_vect)
# True
```