@kylebarron 2021-02-26
One more note: it might be nice to coerce input in these functions to a numpy array? There are some numpy-like types (like some pandas types) that I've found don't work automatically with the current implementation, so a simple `np.array()` on the input could make sense. That would also make it work with arbitrary Python lists.