@ajfriend 2021-03-01
This is great! I'm taking a look now.

> One more note: it might be nice to coerce input in these functions to a numpy array? There are some numpy-like types (like some pandas types) that I've found don't work automatically with the current implementation, so a simple `np.array()` on the input could make sense. That would also make it work with arbitrary Python lists.

Wrapping with `np.array()` makes sense. Do you have a few examples of it failing with pandas types? We could make an issue with these (and the `list` example) to make sure we're handling these cases.

Also, just thinking about long term, I think we'd ultimately like to have the vectorized API use memoryviews so that we don't have a strict dependency on `numpy`. Using `numpy` right now totally makes sense as we're playing around with things, I just want to call out that at some point, we'll want to try to see if we can get memoryviews working.