@ajfriend 2021-03-01
Also, `Py_ssize_t` is a nice catch! I've got to make a note to use that elsewhere in the library.