@ajfriend 2021-03-01
I ran your benchmarks, adding a list comprehension (since they're typically faster than loops), and `np.fromiter`. The list comprehension is a bit faster, but the vectorized version is still the obvious winner.

```python
import h3.api.numpy_int as h3
import h3.unstable.vect as vect
import numpy as np

N = 100_000

lats, longs = np.random.uniform(0, 90, N), np.random.uniform(0, 90, N)
h3_index = vect.geo_to_h3(lats, longs, 9)

vec_fn = np.vectorize(h3.h3_to_parent)

%%timeit
out_loop = []
for i in range(len(h3_index)):
    out_loop.append(h3.h3_to_parent(h3_index[i], 8))
# 101 ms ± 3.2 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

%%timeit
out_list_comp = [h3.h3_to_parent(h, 8) for h in h3_index]
# 88.4 ms ± 4.31 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

%%timeit
out_fromiter = np.fromiter((h3.h3_to_parent(h, 8) for h in h3_index), np.uint64)
# 90.8 ms ± 2.64 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

%%timeit
out_vec_fn = vec_fn(h3_index, 8)
# 76.8 ms ± 3.97 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

%%timeit
out_vect = vect.h3_to_parent(h3_index, 8)
# 557 µs ± 26.4 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```